Optimisations for dealing with sparse data.
After initial scoring we perform a "force connectivity" operation which
does its best to ensure that adjacent elements are connected by ordering
operations.

We also add a local sort operation which moves every element as far left
as we can justify and judiciously apply this. Between the two of these,
this appears to do well on sparse cycles and even improves on some best
cases for some of the tests.